Move swagger2 code in a separate directory
