+++ sanic_openapi/swagger2/spec.py
